pip-requirements.txt
scipy
rdflib
pyparsing==2.1.0
nose
coverage
jinja2
git+git://github.com/gjhiggins/FuXi
kappy
pyzmq
